updates header text to open menu onClick
