+++ src/components/ui/Header/HeaderMenu.tsx
@@ -1,15 +1,48 @@
 import { Menu, Transition } from "@headlessui/react";
+import useAvatar from "../../../hooks/useAvatar";
+import useDisplayName from "../../../hooks/useDisplayName";
+import { useWeb3 } from "../../../web3";
 import DownArrow from "../svg/DownArrow";
+import Avatar from "./Avatar";
 import MenuItems from "./MenuItems";
 
+const ConnectWallet = ({ account }: { account?: string }) => {
+  if (account) {
+    return null;
+  }
+  return <span className="text-sm text-gold-500">Connect Wallet</span>;
+};
+
+const WalletConnected = ({ account }: { account?: string }) => {
+  const accountDisplayName = useDisplayName(account);
+  const avatarURL = useAvatar(account);
+
+  if (!account) {
+    return null;
+  }
+  return (
+    <>
+      <Avatar address={account} url={avatarURL} />
+      <div className="pl-2 flex flex-col items-end">
+        <div className="sm:text-right text-sm text-gold-500">{accountDisplayName}</div>
+      </div>
+    </>
+  );
+};
+
 const HeaderMenu = () => {
+  const { account } = useWeb3();
   return (
     <div className="flex items-center justify-center relative">
       <Menu>
         {({ open }) => (
           <>
-            <Menu.Button className="text-sm font-medium text-white transition duration-150 ease-in-out hover:text-stone-300 focus:outline-none focus:border-blue-300 focus:shadow-outline-blue align-middle">
-              <DownArrow />
+            <Menu.Button className="transition duration-150 ease-in-out hover:text-stone-300 focus:outline-none flex items-center">
+              <ConnectWallet account={account} />
+              <WalletConnected account={account} />
+              <div className="text-sm font-medium text-white hover:text-stone-300 focus:shadow-outline-blue">
+                <DownArrow />
+              </div>
             </Menu.Button>
 
             <Transition
